Refactored project. Controller updates: create/POST working
