+++ home/src/main/webapp/WEB-INF/servConf.xml
@@ -18,6 +18,6 @@
 	<property name="suffix" value=".html"></property>
 </bean>
 
-<context:component-scan base-package="home" />
+<context:component-scan base-package="com.revature" />
 
 </beans>
